#6276: Add missing DEBUG_SANITIZE calls in FW
