+++ tt_metal/hw/inc/debug/watcher_common.h
@@ -9,7 +9,11 @@
 #if defined(WATCHER_ENABLED)
 
 #if defined(COMPILE_FOR_ERISC)
-#include "erisc.h"
+// Forward declare these functions to avoid including erisc.h -> circular dependency via noc_nonblocking_api.h, sanitize_noc.h.
+namespace internal_ {
+void __attribute__((section("code_l1"))) risc_context_switch();
+void disable_erisc_app();
+}
 extern "C" void erisc_early_exit(std::int32_t stack_save_addr);
 #endif
