+++ tt_metal/jit_build/build.cpp
@@ -188,7 +188,6 @@ JitBuildDataMovement::JitBuildDataMovement(const JitBuildEnv& env, int which, bo
     this->defines_ = env_.defines_;
 
     // TODO(pgk): build these once at init into built/libs!
-    this->srcs_.push_back("tt_metal/hw/firmware/src/risc_common.cc");
     this->srcs_.push_back("tt_metal/hw/toolchain/substitutes.cpp");
 
     this->lflags_ = env_.lflags_ + "-Os ";
@@ -379,7 +378,6 @@ JitBuildEthernet::JitBuildEthernet(const JitBuildEnv& env, int which, bool is_fw
         this->includes_ += "-I " + env_.root_ + "tt_metal/hw/firmware/src ";
 
         // TODO(pgk): build these once at init into built/libs!
-        this->srcs_.push_back("tt_metal/hw/firmware/src/risc_common.cc");
         this->srcs_.push_back("tt_metal/hw/toolchain/substitutes.cpp");
         this->srcs_.push_back("tt_metal/hw/firmware/src/" + env_.aliased_arch_name_ + "/noc.c");
         if (this->is_fw_) {
